fixup! refactor: extract in-memory tests
